requirements.txt
# Production (pip): use Python 3.10+ when possible. On Python 3.9, pydantic-settings is
# capped at 2.11.x (2.12+ requires newer Python). Dev/uv uses pyproject.toml (3.12+).
# Prefer: uv sync --locked. If you use pip, install gunicorn separately.
asyncpg>=0.31.0
bcrypt==4.1.2
fastapi==0.104.1
gunicorn>=23.0.0
fastapi-cache2[redis]>=0.2.2
google-auth==2.23.4
httpx>=0.24.0,<0.26.0
mangum>=0.21.0
pydantic-settings>=2.10.0,<2.12
pydantic[email]>=2.10.0,<3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
redis>=4.2.0rc1,<5.0.0
reportlab>=4.0.0
requests==2.31.0
sendgrid>=6.11.0
slowapi>=0.1.9
sqlalchemy==2.0.23
starlette>=0.27.0
stellar-sdk>=11.0.0
supabase==2.3.4
ujson>=5.9.0
uvicorn[standard]==0.24.0
